requirements.txt
azure-storage-blob
pandas
scikit-learn
sentence-transformers
openai>=1.0.0
crawl4ai>=0.6.3
requests
beautifulsoup4
python-dotenv
playwright
tenacity
fastapi
uvicorn
promptflow




